EDIT: int WorldTime, smooth lerp Skybox properties; ADD: Daylight cycle

+++ Assets/Scripts (User)/IWorldTimeListener.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IWorldTimeListener
 {
-    public void OnTimeChanged(float currentTime)
+    public void OnTimeChanged(int currentTime)
     {
 
     }
